Move tests to be under test/unit
