+++ test/unit/test_executor_context_set_emits.py
